refactor: object condition parsing and expression formatting
- object condition parsing => src/query_object.js
- expression formatting => src/expr_formatter.js

currently the `spell` object is still passed along the `formatXxx()` functions in expr_formatter.js. There might be two ways to refactor this:

- Turn `formatXxx()` into `ExprFormatter()` instances with `this.spell`;
- Turn expr objects into `Expr`, `OpExpr`, etc. with `this.spell`;

both requires lots of work but without little perf improvements, hence postponded further.
