+++ src/drivers/abstract/spellbook.js
@@ -2,25 +2,8 @@
 
 const SqlString = require('sqlstring');
 
-const { precedes, copyExpr, findExpr, walkExpr } = require('../../expr');
-
-/**
- * Find model by qualifiers.
- * @example
- * findModel(spell, ['comments'])
- * findModel(spell)
- *
- * @param {Spell} spell
- * @param {string[]} qualifiers
- */
-function findModel(spell, qualifiers) {
-  const qualifier = qualifiers && qualifiers[0];
-  const Model = qualifier && qualifier != spell.Model.tableAlias
-    ? (spell.joins.hasOwnProperty(qualifier) ? spell.joins[qualifier].Model : null)
-    : spell.Model;
-  if (!Model) throw new Error(`Unabled to find model ${qualifiers}`);
-  return Model;
-}
+const { copyExpr, findExpr, walkExpr } = require('../../expr');
+const { formatExpr, formatConditions, collectLiteral } = require('../../expr_formatter');
 
 /**
  * Format orders into ORDER BY clause in SQL
@@ -29,183 +12,9 @@ function findModel(spell, qualifiers) {
  */
 function formatOrders(spell, orders) {
   return orders.map(([token, order]) => {
-    const column = formatColumn(spell, token);
+    const column = formatExpr(spell, token);
     return order == 'desc' ? `${column} DESC` : column;
   });
-}
-
-/**
- * Format token into identifiers/functions/etc. in SQL
- * @example
- * formatColumn(spell, { type: 'id', value: 'title' })
- * // => `title`
- *
- * formatColumn(spell, { type: 'func', name: 'year', args: [ { type: 'id', value: 'createdAt' } ] })
- * // => YEAR(`createdAt`)
- *
- * @param {Spell}  spell
- * @param {Object} token
- */
-function formatColumn(spell, token) {
-  if (token.type == 'id') {
-    return formatIdentifier(spell, token);
-  } else {
-    return formatExpr(spell, token);
-  }
-}
-
-/**
- * Format identifiers into escaped string with qualifiers.
- * @param {Spell}  spell
- * @param {Object} ast
- */
-function formatIdentifier(spell, ast) {
-  const { value, qualifiers } = ast;
-  const Model = findModel(spell, qualifiers);
-  const column = Model.unalias(value);
-  const { escapeId } = spell.Model.driver;
-
-  if (qualifiers && qualifiers.length > 0) {
-    return `${qualifiers.map(escapeId).join('.')}.${escapeId(column)}`;
-  } else {
-    return escapeId(column);
-  }
-}
-
-const extractFieldNames = ['year', 'month', 'day'];
-
-function formatFuncExpr(spell, ast) {
-  const { name, args } = ast;
-  const { type } = spell.Model.driver;
-
-  // https://www.postgresql.org/docs/9.1/static/functions-datetime.html
-  if (type === 'postgres' && extractFieldNames.includes(name)) {
-    return `EXTRACT(${name.toUpperCase()} FROM ${args.map(arg => formatExpr(spell, arg)).join(', ')})`;
-  }
-
-  return `${name.toUpperCase()}(${args.map(arg => formatExpr(spell, arg)).join(', ')})`;
-}
-
-/**
- * The `... IS NULL` predicate is not parameterizable.
- * - https://github.com/brianc/node-postgres/issues/1751
- * @param {Array} values the collected values
- * @param {Object} ast the abstract syntax tree
- * @returns {Array} values
- */
-function collectLiteral(values, ast) {
-  walkExpr(ast, ({ type, value }) => {
-    if (type == 'literal' && value != null) {
-      if (Array.isArray(value)) {
-        values.push(...value);
-      } else {
-        values.push(value);
-      }
-    }
-  });
-  return values;
-}
-
-function formatLiteral(spell, ast) {
-  const { value } = ast;
-
-  if (value == null) {
-    return 'NULL';
-  } else if (Array.isArray(value)) {
-    if (value.length) return `(${value.map(() => '?').join(', ')})`;
-    return '(NULL)';
-  } else {
-    return '?';
-  }
-}
-
-/**
- * Format the abstract syntax tree of an expression into escaped string.
- * @param {Spell}  spell
- * @param {Object} ast
- */
-function formatExpr(spell, ast) {
-  const { type, name, value, args } = ast;
-  switch (type) {
-    case 'literal':
-      return formatLiteral(spell, ast);
-    case 'subquery':
-      return `(${value.toSqlString()})`;
-    case 'wildcard':
-      return '*';
-    case 'alias':
-      return `${formatExpr(spell, args[0])} AS ${formatIdentifier(spell, ast)}`;
-    case 'mod':
-      return `${name.to.toUpperCase()} ${formatExpr(spell, args[0])}`;
-    case 'id':
-      return formatIdentifier(spell, ast);
-    case 'op':
-      return formatOpExpr(spell, ast);
-    case 'func':
-      return formatFuncExpr(spell, ast);
-    case 'raw':
-      // return value directly
-      return value;
-    default:
-      throw new Error(`Unexpected type ${type}`);
-  }
-}
-
-/**
- * Check if current token is logical operator or not, e.g. `AND`/`NOT`/`OR`.
- * @param {Object} ast
- */
-function isLogicalOp({ type, name }) {
-  return type == 'op' && ['and', 'not', 'or'].includes(name);
-}
-
-/**
- * Format `{ type: 'op' }` expressions into escaped string.
- * @param {Spell}  spell
- * @param {Object} ast
- */
-function formatOpExpr(spell, ast) {
-  const { name, args } = ast;
-  const params = args.map(arg => {
-    return isLogicalOp(ast) && isLogicalOp(arg) && precedes(name, arg.name) < 0
-      ? `(${formatExpr(spell, arg)})`
-      : formatExpr(spell, arg);
-  });
-
-  if (name == 'between' || name == 'not between') {
-    return `${params[0]} ${name.toUpperCase()} ${params[1]} AND ${params[2]}`;
-  }
-  else if (name == 'not') {
-    return `NOT ${params[0]}`;
-  }
-  else if ('!~-'.includes(name) && params.length == 1) {
-    return `${name} ${params[0]}`;
-  }
-  else if (args[1].type == 'literal' && args[1].value == null && !isLogicalOp(ast)) {
-    if (['=', '!='].includes(name)) {
-      const op = name == '=' ? 'IS' : 'IS NOT';
-      return `${params[0]} ${op} NULL`;
-    } else {
-      throw new Error(`Invalid operator ${name} against null`);
-    }
-  }
-  // IN (1, 2, 3)
-  // IN (SELECT user_id FROM group_users)
-  else if ((args[1].type == 'literal' && Array.isArray(args[1].value)) || args[1].type == 'subquery') {
-    let op = name;
-    if (name == '=') {
-      op = 'in';
-    } else if (name == '!=') {
-      op = 'not in';
-    }
-    if (['in', 'not in'].includes(op)) {
-      return `${params[0]} ${op.toUpperCase()} ${params[1]}`;
-    } else {
-      throw new Error(`Invalid operator ${name} against ${args[1].value}`);
-    }
-  } else if (params[1] !== '') {
-    return `${params[0]} ${name.toUpperCase()} ${params[1]}`;
-  }
 }
 
 /**
@@ -228,7 +37,7 @@ function formatSelectWithoutJoin(spell) {
     columns.reduce(collectLiteral, values);
     const selects = [];
     for (const token of columns) {
-      const column = formatColumn(spell, token);
+      const column = formatExpr(spell, token);
       if (!selects.includes(column)) selects.push(column);
     }
     chunks.push(`${selects.join(', ')}`);
@@ -254,7 +63,7 @@ function formatSelectWithoutJoin(spell) {
   }
 
   if (groups.length > 0) {
-    const groupColumns = groups.map(group => formatColumn(spell, group));
+    const groupColumns = groups.map(group => formatExpr(spell, group));
     chunks.push(`GROUP BY ${groupColumns.join(', ')}`);
   }
 
@@ -371,7 +180,7 @@ function formatSelectExpr(spell, values) {
 
   for (const token of columns) {
     collectLiteral(values, token);
-    const selectExpr = formatColumn(spell, token);
+    const selectExpr = formatExpr(spell, token);
     const qualifier = token.qualifiers ? token.qualifiers[0] : '';
     const list = map[qualifier] || (map[qualifier] = []);
     list.push(selectExpr);
@@ -444,7 +253,7 @@ function formatSelectWithJoin(spell) {
   }
 
   if (groups.length > 0) {
-    chunks.push(`GROUP BY ${groups.map(group => formatColumn(spell, group)).join(', ')}`);
+    chunks.push(`GROUP BY ${groups.map(group => formatExpr(spell, group)).join(', ')}`);
   }
 
   if (havingConditions.length > 0) {
@@ -511,22 +320,6 @@ function formatDelete(spell) {
   } else {
     return { sql: chunks.join(' ') };
   }
-}
-
-/**
- * Format an array of conditions into an expression. Conditions will be joined with `AND`.
- * @param {Object[]} conditions - An array of parsed where/having/on conditions
- */
-function formatConditions(spell, conditions) {
-  return conditions
-    .map(condition => {
-      return isLogicalOp(condition) && condition.name == 'or' && conditions.length > 1
-        ? `(${formatExpr(spell, condition)})`
-        : formatExpr(spell, condition);
-    })
-    // filter empty condition
-    .filter((condition) => !!condition)
-    .join(' AND ');
 }
 
 /**
